ML-Interview-Preparation/MlOps-Project/End-to-End-Mlops-Image-Classification-Project/Frontend/requirements.txt
streamlit
pillow
requests
